LayoutTests/imported/w3c/web-platform-tests/workers/semantics/structured-clone/shared-expected.txt

CONSOLE MESSAGE: line 19: ReferenceError: Can't find variable: SharedWorker

Harness Error (FAIL), message = ReferenceError: Can't find variable: SharedWorker
